LayoutTests/fast/text-extraction/avoid-extracting-password-fields-expected.txt


(ROOT
  (TEXT "\n1. Password field\n")
  (TEXT "" {editable (secure)})
  (TEXT "\n2. Text field (previously a password field)\n")
  (TEXT "" {editable (secure)})
  (TEXT "\n3. Password field (previously a text field)\n")
  (TEXT "" {editable (secure)})
  (TEXT "\n4. Plain text field\n")
  (TEXT "fourth\n" {editable}))
